# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AKSIS Core API
  version: 1.13.0
  description: >-
    Integrated Phase 03-15 REST contract covering authentication and tenant
    context, schools/classes/students, academic periods and student history,
    card lifecycle, and secure device provisioning/runtime operations.
  x-aksis-phases:
    - phase: 3
      resources: [Auth, Schools, Classes, Students]
    - phase: 4
      resources: [Academic Years, Student History, Cards]
    - phase: 5
      resources: [Devices]
    - phase: 6
      resources: [Gate Attendance, Offline Sync]
    - phase: 7
      resources: [Admin Context, Attendance Read Model, Device Inventory]
    - phase: 8
      resources: [Parent Linking, Children, Parent Timeline]
    - phase: 9
      resources: [Waste Transactions, Ranking, Summary]
    - phase: 10
      resources: [Extracurriculars, Sessions, Members, Attendance]
    - phase: 11
      resources: [Library Visits, Offline Sync, Class Summary]
    - phase: 12
      resources: [LED Content, Gateway State, Heartbeat, Acknowledgement]
    - phase: 13
      resources: [Command Center, Today Aggregate]
    - phase: 14
      resources: [Card Writer Jobs, Station Lease, Read-back Verification]
    - phase: 15
      resources: [Reporting, Audit, Monitoring]
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
    schoolId: []
paths:
...
  /reports/{domain}:
    get:
      tags: [Reporting]
      operationId: exportTenantReport
      summary: Export a tenant-scoped operational report
      parameters:
        - { name: domain, in: path, required: true, schema: { type: string, enum: [attendance, waste, library, extracurricular] } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: format, in: query, schema: { type: string, enum: [json, csv], default: json } }
      responses: { '200': { description: JSON rows or downloadable CSV }, '403': { $ref: '#/components/responses/Forbidden' } }
  /reports/audit/logs:
    get:
      tags: [Reporting]
      operationId: listAuditLogs
      summary: List append-only privileged action records
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses: { '200': { description: Paginated audit records }, '403': { $ref: '#/components/responses/Forbidden' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    schoolId:
      type: apiKey
      in: header
      name: X-School-Id
      description: UUID of the active school membership.
    deviceAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Per-device credential formatted as `Device <token>`.
    deviceId:
      type: apiKey
      in: header
      name: X-Device-Id
      description: Registered device UUID; school context is derived server-side.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
...
    Forbidden:
      description: Tenant membership or permission is missing
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error, request_id]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [AUTH_REQUIRED, AUTH_INVALID, INVALID_CREDENTIALS, SESSION_REFRESH_FAILED, FORBIDDEN, TENANT_REQUIRED, SCHOOL_NOT_FOUND, ACADEMIC_YEAR_NOT_FOUND, CLASS_NOT_FOUND, STUDENT_NOT_FOUND, CARD_NOT_FOUND, DEVICE_NOT_REGISTERED, DEVICE_AUTH_INVALID, DEVICE_NOT_GATE, ATTENDANCE_DISABLED, CARD_BLOCKED, STUDENT_INACTIVE, GATE_RULE_VIOLATION, PARENT_LINK_INVALID, CHILD_NOT_LINKED, RESOURCE_NOT_FOUND, INVALID_JSON, PAYLOAD_TOO_LARGE, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, ROUTE_NOT_FOUND, DATABASE_ERROR, SERVER_UNAVAILABLE, INTERNAL_ERROR]
            message: { type: string }
            details: {}
        request_id: { type: string }
...